Finished porting `realtime.inc` to C, small fixes.
* font.c: Initial release.
* font.h: Initial release.
* font.pas: Initial release.
* go32/mouse.c: Initial release.
* go32/mouse.pas: Initial release.
* realtime/pas/_decay_bar_str.pas: Moved into "ui/pas/_decay_bar_str.pas".
* realtime/pas/decay_bars_refresh.pas: Moved into "ui/pas/decay_bars_refresh.pas".
* realtime/pas/show_bar.pas: Moved into "ui/pas/show_bar.pas".
* realtime/decay_bars_refresh.c: Moved into "ui/decay_bars_refresh.c".
* realtime/show_bar.c: Moved into "ui/decay_bars_window.c" as "decay_bars_window__draw_bar" routine.
* realtime.inc: Finished porting to C.
* realtime.c: Finished porting to from Pascal.
* realtime.h: Finished porting to from Pascal.
* ui/pas/_decay_bar_str.pas: Moved from "realtime/pas/_decay_bar_str.pas".
* ui/pas/_vol_bar_str.pas: Initial release (saved from "realtime.inc").
* ui/pas/decay_bars_refresh.pas: Moved from "realtime/pas/decay_bars_refresh.pas".
* ui/pas/show_bar.pas: Moved from "realtime/pas/show_bar.pas".
* ui/pas/status_refresh.pas: Initial release (saved from "realtime.inc").
* ui/pas/trace_update_proc.pas: Initial release (saved from "realtime.inc").
* ui/_is_status_window_visible.c: Initial release (ported from "realtime.inc").
* ui/_time_playing_str.c: Initial release (ported from "realtime.inc").
* ui/decay_bars_refresh.c: Moved from "realtime/decay_bars_refresh.c", finished porting from Pascal.
* ui/decay_bars_window.c: Initial release (ported from "realtime.inc").
* ui/decay_bars_window.h: Initial release (ported from "realtime.inc").
* ui/main_window.c: Initial release (ported from "realtime.inc").
* ui/main_window.h: Initial release (ported from "realtime.inc").
* ui/pattern_editor_window.c: Initial release (ported from "realtime.inc").
* ui/pattern_editor_window.h: Initial release (ported from "realtime.inc").
* ui/status_refresh.c: Initial release (ported from "ui/pas/status_refresh.pas").
* ui/status_window.c: Initial release (ported from "realtime.inc").
* ui/status_window.h: Initial release (ported from "realtime.inc").
* ui/trace_update_proc.c: Initial release (ported from "ui/pas/trace_update_proc.pas").
* adt2unit.pas: Ported "decay_bars_refresh", "status_refresh" and "trace_update_proc" routines to C.
* adt2unit.c: Added "decay_bars_refresh", "status_refresh" and "trace_update_proc" routines (ported from "realtime.inc").
* adt2unit.h: Updated due to changes in "adt2unit.c".
* stringio/Bpm2str.c: Replaced "float" type with "double".
* stringio/ExpCStrR.c: Initial release.
* stringio.c: Added "ExpCStrR" routine.
* stringio.h: Updated due to changes in "stringio.c" and included routines.
* stringio.pas: Updated due to changes in "stringio.h".
* txtscrio/VScrollBar.c: Fixed scroll bar drawing.
* README.md: Added Ralf Brown's Interrupt List (RBIL) reference.
* Makefile: Updated.

+++ stringio.pas
@@ -71,7 +71,7 @@ function FilterStr2(str: String; var chr0: tCHARSET; chr1: Char): String; cdecl;
 function Num2str(num: Longint; base: Byte): String; cdecl; external;
 function Str2num(str: String; base: Byte): Longint; cdecl; external;
 {$IFNDEF ADT2PLAY}
-function Bpm2str(bpm: Single): String; cdecl; external;
+function Bpm2str(bpm: Double): String; cdecl; external;
 {$ENDIF} // NOT DEFINED(ADT2PLAY)
 
 {$IFNDEF ADT2PLAY}
